refactor(agent-ux): split 1303-line agent-ux-tools.ts into focused modules
agent-ux-tools.ts had grown into a grab bag of 13 tools, 12 helper
functions, and 2 static migration maps — 1303 lines of mixed
concerns. Split into a sub-module group under tools/agent-ux/ with
each file owning one cohesive area:

  agent-ux/shared.ts          (228) — pure helpers reused across the group
  agent-ux/migration-tools.ts (183) — stdlib + builtin migration data
                                      and 3 lookup tools
  agent-ux/edit-tools.ts      (311) — single-file edits (apply_rename,
                                      add_missing_clauses,
                                      infer_fixity_conflicts) + the
                                      triage_error helper they share
  agent-ux/import-tools.ts    (247) — import suggestion + clash-source
                                      resolution (both walk the import
                                      graph)
  agent-ux/options-tools.ts   (220) — effective_options +
                                      project_config introspection
  agent-ux/project-tools.ts   (338) — project-wide scans
                                      (postulate_closure,
                                      project_progress, bulk_status)

The original `agent-ux-tools.ts` is now a 38-line barrel that delegates
to each sub-module's `register*Tools(server, session, repoRoot)`, so
existing imports in `register-core-tools.ts` and the unit-test suite
keep working unchanged. Every sub-module is well under the 500-line
ceiling so future tool additions land in the cohesive home rather
than re-inflating one giant file.

No behavior change: 1186 vitest cases pass identically.
